[BACKLOG-5864] - REG: PMR: Reducer fails on the client due to job running.
